Renamed to actionTypes and removed default export
